+++ src/main/kotlin/model/NdcBounds.kt
@@ -4,7 +4,7 @@ import com.midnightcrowing.gui.base.Window
 
 
 // NDC坐标系的边界
-data class NdcBounds(val left: Float, val top: Float, val right: Float, val bottom: Float) {
+data class NdcBounds(val left: Double, val top: Double, val right: Double, val bottom: Double) {
     /**
      * 将 NDC (Normalized Device Coordinates) 归一化坐标转换为屏幕坐标。
      *
@@ -13,7 +13,7 @@ data class NdcBounds(val left: Float, val top: Float, val right: Float, val bott
      * @param y NDC 坐标系中的 Y 坐标，范围为 [-1, 1]
      * @return 转换后的屏幕坐标 (screenX, screenY)，范围为 [0, window.width] 和 [0, window.height]
      */
-    private fun convertNdcToScreen(window: Window, x: Float, y: Float): Point {
+    private fun convertNdcToScreen(window: Window, x: Double, y: Double): Point {
         val screenX = (x + 1) * (window.width / 2f)
         val screenY = (-y + 1) * (window.height / 2f) // OpenGL NDC Y 轴向上，屏幕 Y 轴向下
         return Point(screenX, screenY)
