+++ src/main/kotlin/model/Point.kt
@@ -1,4 +1,4 @@
 package com.midnightcrowing.model
 
 
-data class Point(val x: Float, val y: Float)
+data class Point(val x: Double, val y: Double)
