+++ src/main/kotlin/utils/FontSizeUtil.kt
@@ -3,14 +3,14 @@ package com.midnightcrowing.utils
 object FontSizeUtil {
     fun calculateFontSize(
         windowWidth: Int,
-        minFontSize: Float = 16f, maxFontSize: Float = 24f,
+        minFontSize: Double = 16.0, maxFontSize: Double = 24.0,
         minWindowWidth: Int = 1216, maxWindowWidth: Int = 1920,
-    ): Float {
+    ): Double {
         return when {
             windowWidth <= minWindowWidth -> minFontSize
             windowWidth >= maxWindowWidth -> maxFontSize
             else -> {
-                val scale = (windowWidth - minWindowWidth).toFloat() / (maxWindowWidth - minWindowWidth)
+                val scale = (windowWidth - minWindowWidth) / (maxWindowWidth - minWindowWidth)
                 minFontSize + scale * (maxFontSize - minFontSize)
             }
         }
